fix(dashboard): split bff and user_store modules for file-length gate
Extract admin/agent BFF routes and UserStore profile ops into dedicated
modules so SLOC stays under the 300-line quality gate.
